feat: Bookshelf AuthorWorksPage (Find-1, part 1) + online-layer scope
First slice of the Find rework's destinations. Replaces AuthorBooksPage with
AuthorWorksPage — token-based + dark (closes the F3 straggler for this
surface), reached from the existing author search. v1 shows owned works only;
the "Missing" scope (works you don't own) needs the author's online
bibliography and moves to the online-layer arc.

- Theming/Covers.cs: shared async cover loader (EnsureCoverCachedAsync →
  swap placeholder for image), de-duping the per-page pattern.
- Pages/AuthorWorksPage: owned works by canonical author (alias rollup),
  BookSnapshot rows with cover + "★★★☆☆ · Status", tap → web book detail.
  All colours via tokens/SetThemeColor → dark mode.
- AuthorSearchPage now pushes AuthorWorksPage; AuthorBooksPage deleted.

Scope recorded (Drew's 2026-06-16 call — online is its own arc):
- TODO #54: Bookshelf online layer (work-level ownership · author Missing ·
  other editions).
- Redesign brief §10: Recent-lookups list deferred (decide this arc or push);
  online layer pointer.

Builds (MAUI Android, 0 warnings). Next Find-1 slice: ResultPage +
WorkEditionPanel (the scan/ISBN destination).
