--- BookTracker.Mobile/Pages/AuthorBooksPage.xaml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="BookTracker.Mobile.Pages.AuthorBooksPage"
-             BackgroundColor="#FAF6EC">
-
-    <!-- AuthorBooksPage — second leaf of the author lookup flow.
-         Renders all books credited to the canonical author (which
-         includes books credited to any alias that rolls up to that
-         canonical via Author.CanonicalId).
-
-         Title is set in code-behind to "Books by {AuthorName}" once
-         the author is loaded; the system nav bar shows that + the
-         back arrow. -->
-
-    <Grid RowDefinitions="Auto,*" Padding="20,16,20,0">
-
-        <Label x:Name="HeaderLabel"
-               Grid.Row="0"
-               FontSize="14"
-               TextColor="#6B5D4A"
-               LineBreakMode="WordWrap" />
-
-        <ScrollView Grid.Row="1" Margin="0,12,0,0">
-            <VerticalStackLayout x:Name="BooksLayout"
-                                 Spacing="8"
-                                 Padding="0,0,0,20" />
-        </ScrollView>
-    </Grid>
-</ContentPage>
