+++ BookTracker.Mobile/Pages/AuthorWorksPage.xaml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="BookTracker.Mobile.Pages.AuthorWorksPage"
+             Style="{StaticResource Page}">
+
+    <!-- Author drill-down: the works you own by the canonical author (alias
+         credits roll up). Title is set in code-behind. v1 is owned-only — the
+         "Missing" scope (works you don't own) needs the author's online
+         bibliography, deferred to the online-layer arc (TODO #54). -->
+
+    <Grid RowDefinitions="Auto,*" Padding="20,16,20,0">
+        <Label x:Name="HeaderLabel"
+               Grid.Row="0"
+               FontSize="14"
+               LineBreakMode="WordWrap"
+               TextColor="{AppThemeBinding Light={StaticResource TextMutedL}, Dark={StaticResource TextMutedD}}" />
+
+        <ScrollView Grid.Row="1" Margin="0,12,0,0">
+            <VerticalStackLayout x:Name="WorksLayout"
+                                 Spacing="8"
+                                 Padding="0,0,0,20" />
+        </ScrollView>
+    </Grid>
+</ContentPage>
