+++ TODO.md
@@ -8,6 +8,7 @@ Size legend: **XS** ≈ one-line fix, **S** ≈ one short session, **M** ≈ a d
 
 | # | Category | Name | Description | Size |
 |---|---|---|---|---|
+| 54 | Mobile | Bookshelf online layer — "do better when online" (separate enhancement arc) | The Bookshelf cache holds only the *owned* library, so the redesign's v1 Find is offline search/scan of what you own. Three features need online data and are deferred to their own arc (Drew's 2026-06-16 call): (a) **work-level ownership** — a scanned ISBN that's a different printing of a work you own should read "you have this work, not this edition" (needs ISBN→work resolution against the web catalogue); (b) **AuthorWorks "Missing only"** — show works by an author you *don't* own (needs the author's full bibliography, online); (c) **"Other editions of this work"** — the disabled online-only row in `WorkEditionPanel`. All must degrade gracefully offline and never block the primary owned-library answer. Pairs with the parked items in `docs/BOOKSHELF-UI-REDESIGN.md` §10. | M |
 | 1 | Infra / Security | Bacpac sync from prod → staging | Empty staging catches schema syntax errors but **misses every data-shape failure**: `NOT NULL` against existing nulls, unique-index/constraint additions where data has duplicates, FK additions where data has orphans, CHECK violations, type-conversion failures, perf-on-real-data timeouts. EF migrations on SQL Server are transactional so prod can't *corrupt* — but the deploy still fails after merge. Add an on-demand prod → staging bacpac sync (sibling of the existing `infra/refresh-local-db.ps1` for prod → local) — either a `refresh-staging-db.ps1` script or a GH Action workflow (`sync-staging-data.yml`) so it's a one-click pre-flight before any review-tag migration. Workflow path is preferred because it doesn't need SqlPackage on the dev box and runs from inside the VNet. | S |
 | 2 | Security | Nonce-based CSP middleware | Replace the current meta-tag CSP's `'unsafe-inline'` / `'unsafe-eval'` on `script-src` with a per-request nonce threaded through Blazor's framework inlines. The meta-tag CSP added in the initial audit (SECURITY-AUDIT.md §4) is a baseline — this is the tightening. Real XSS hardening; moderate implementation work because Blazor Server's script injection points need to read the nonce. | M |
 | 4 | Meta | Package reusable Claude prompts into a `prompts/` directory | Two reusable paste-into-another-Claude prompts were extracted during development (git-history-rewrite, going-public walkthrough). Text is saved locally by Drew in separate files but not in this repo. Add a `prompts/` directory at repo root, one `.md` per prompt, with a small `prompts/README.md` explaining the "reusable-instructions-as-deliverable" shape (see `retros/retro_reusable_instructions.md`). Link from the main README. More prompts will accrue over time — set up the pattern once, populate as it grows. **Note:** a "skill chassis for guided procedures" (mutating workflows like the going-public walkthrough) is a different shape from the audit-skill chassis (read-only analysis) and would warrant its own pilot if the prompts list grows past 3–4 candidates. For now, the simple `prompts/` directory is the right answer. | S |
