+++ docs/BOOKSHELF-UI-REDESIGN.md
@@ -354,6 +354,8 @@ Motion.Enabled = scale > 0f;
 | **Sync mode on start** | **Decided (2026-06-16): auto-sync silently** | When online at launch, fire the delta sync in the background — no prompt. Wire in PR4. |
 | **Wishlist default sort** | **Decided (2026-06-16): Priority default** | Keep High/Medium/Low grouping as default; offer Author/Series as alternate sorts. PR4. |
 | **Gaps tab longevity** | **Watch** | Candidate for later deprecation as missing→wishlist grows. Keep self-contained. |
+| **Recent-lookups list** (§4.1) | **Deferred from v1 — decide this arc or push (2026-06-16)** | Find opens to an empty search-ready state for now. Adding it needs a lookup-history store (in-memory resets on restart; persisted = new SQLite table). Decide before the arc closes whether it earns a slot or moves to TODO. |
+| **Online layer (work-level ownership · author Missing · other editions)** | **Deferred to its own enhancement arc — TODO #54 (2026-06-16)** | The cache holds only the owned library, so all "what am I *missing*" features need online catalogue data. v1 Find is offline search/scan of owned items; the online layer is a separate arc. |
 | **Transitional stragglers — recheck at arc end** | **Accepted as transitional (2026-06-16 arc review F3/F4/F5)** | (F3) `MainPage`/`ScanPage`/Author/Title still light-only inline hex → dark mode looks broken on the default Find tab + pushed search pages until PR3 rebuilds them. (F4) `KillerButton`/`PrimaryButton`/`OutlinedButton`/`GapProgress` styles + (F5) ~24 tokens and 13/16 `TablerIcons` glyphs have no consumers yet (defined ahead of the surfaces that use them; icon codepoints unverified-until-rendered). **All deliberately deferred** — the surfaces that consume them are built in PR3–5. **At end of arc, recheck:** any page still on inline hex, any style/token/glyph still unconsumed → migrate or delete. Don't let the foundation's unused half calcify. |
 
 ---
